Remove an infinite loop in worker using new Resque gem.
Changed "RESERVE_ARG" variable to "RESERVE_ACCEPTS_INTERVAL" to make
the intent of the code clearer and moved the sleep(interval) to the
branch which isn't handled by the reserve(interval) call.

Tested on Mac OS X with Trinidad 1.4.4, JRuby 1.7.2, JDK 1.7.0_u10 with
the worker in VVERBOSE mode. The log shows the worker sleeping for the
configured interval (10 seconds) and the CPU usage is normal.